+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Web search: Added support for Grok as an internal search provider.
 - Google: Set `thought=True` on content when replaying `ContentReasoning` back to the model.
 - Transcript: Add globally unique `uuid` field and `metadata` field to `Event`.
+- Transcript: Add `message_id` field to `ToolEvent` for corresponding `ChatMessageTool`.
 - Eval log: Add option to select sample by `uuid` in `read_eval_log_sample()`.
 
 ## 0.3.112 (03 July 2025)